pyproject.toml
[tool.poetry]
name = "AgentZer0"
version = "0.1.0"
description = "A Discord bot that uses an LLM with the Model Context Protocol (MCP)"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "AgentZer0"}]

[tool.poetry.dependencies]
python = "^3.8.1"
discord-py = "^2.3.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.5"
groq = "^0.4.0"
ujson = "^5.7.0"
pynacl = {version = "^1.5.0", optional = true}
httpx = "0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "discord_mcp_bot.main:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"